Moved InternalExecute to own internal class
